requirements.txt
pandas
scikit-learn
xgboost
joblib
umap-learn
tensorflow
librosa
streamlit
tqdm
numpy
